+++ src/components/CardPeople/cardPeople.module.css
